# docs/openapi.yaml
openapi: 3.0.3
info:
  title: InspeXO Auth & Admin API
  version: 1.11.0
  description: |
    Backend authentication, authorization, verification, and admin
    management for InspeXO. This document is generated by hand and kept
    honest by `tests/auth/openapi.contract.test.ts`, which asserts every
    route file has a matching entry here and vice versa.

    ## UI integration notes (read first)

    * There is NO public registration. No `/register` endpoint exists and
      none is documented here. Users are provisioned by ADMIN only.
    * The login page (`/auth/login`) currently offers Microsoft SSO only
      (Entra is not wired). There is NO email/password form and NO
      forgot-password screen yet. The `POST /api/auth/*` contract below
      stands alone; the reset/verify links target future pages at
      `/auth/reset-password?token=…` and `/auth/verify-email?token=…`,
      which do not exist yet.
    * The Admin Dashboard UI does not exist yet. It must consume
      `/api/admin/*` over HTTP and must never import the DAL/database.

    ## Authentication model

    * Sessions are PASETO v3.local tokens (`sub: userId`, `jti`, 8 h
      expiry) carried ONLY in the `__Host-inspexo_session` cookie
      (`HttpOnly`, `Secure`, `SameSite=Lax`, `Path=/`). Never
      localStorage. The token carries identity only — role, department,
      and site assignments are re-read from the database per request.
    * Every request is validated twice: cryptographic expiry in the
...

...

...

...

...
servers:
...
paths:
...
  /api/auth/forgot-password:
    post:
      tags: [auth]
      summary: Request a password-reset link
      description: |
        ALWAYS returns 202 with `{sent: true}` — existent or not,
        active or not, throttled or not, malformed or not. Tokens are
        issued only for existing active users (opaque random, SHA-256
        hash stored, 1 h expiry, single-use, previous tokens
        superseded) and emailed via Resend. Throttled per email
        (5 / 15 min) without changing the status.
      security: []
      parameters:
        - $ref: '#/components/parameters/OriginHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  maxLength: 255
      responses:
        '202':
          description: Accepted (always, regardless of outcome)
          content:
            application/json:
              schema:
                type: object
                required: [sent]
                properties:
                  sent:
                    type: boolean
                    example: true
        '403':
          $ref: '#/components/responses/ForbiddenOrigin'
  /api/auth/reset-password:
    post:
      tags: [auth]
      summary: Redeem a reset token with a new password
      description: |
        Anonymous like verification. Enforces the 12–1024 policy, then
        atomically: marks the token used, stores the Argon2id hash,
        clears `mustChangePassword` (self-chosen secret), stamps the
        credential time, and revokes EVERY session — no session is
        trusted after a reset. Reuse/expiry/garbage collapse to one
        generic 400. Responses never contain secrets.
      security: []
      parameters:
        - $ref: '#/components/parameters/OriginHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, newPassword]
              properties:
                token:
                  type: string
                  minLength: 1
                  maxLength: 256
                newPassword:
                  type: string
                  minLength: 12
                  maxLength: 1024
      responses:
        '200':
          description: Password reset; all sessions revoked
          content:
            application/json:
              schema:
                type: object
                required: [reset]
                properties:
                  reset:
                    type: boolean
                    example: true
        '400':
          description: Invalid, expired, or already-used token (generic message)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              example:
                error: Invalid or expired reset link.
        '403':
          $ref: '#/components/responses/ForbiddenOrigin'
        '422':
          $ref: '#/components/responses/InvalidRequest'
...
components:
  parameters:
    OriginHeader:
      name: Origin
      in: header
      description: Must match the request host on mutating calls (CSRF defence in depth alongside SameSite=Lax)
      schema:
        type: string
...
  schemas:
...
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Generic, safe-to-display message
        code:
          type: string
          description: Stable machine code (e.g. FORBIDDEN_SITE)
          example: FORBIDDEN_SITE
  responses:
...
    ForbiddenOrigin:
      description: Mutating request failed the same-origin check
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
          example:
            error: Forbidden.
    InvalidRequest:
      description: Malformed body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
          example:
            error: Invalid request.
...